refactor(project structure):  adiciona a todos os apps ao seu nome '_app' e altera os campos que possuiam mencao a cada um deles
