+++ projetos_app/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class TarefasConfig(AppConfig):
+class ProjetosConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'tarefas'
+    name = 'projetos_app'
