+++ usuarios_app/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class ProjetosConfig(AppConfig):
+class UsuariosConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'projetos'
+    name = 'usuarios_app'
